basic cplug project

+++ synth/src/fm_algo.c
@@ -1,5 +1,5 @@
 #include "fm_algo.h"
-float fm_algo0000(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0000(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -18,7 +18,7 @@ float fm_algo0000(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output;
 }
-float fm_algo0100(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0100(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -37,7 +37,7 @@ float fm_algo0100(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output;
 }
-float fm_algo0200(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0200(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -56,7 +56,7 @@ float fm_algo0200(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output;
 }
-float fm_algo0300(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0300(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -75,7 +75,7 @@ float fm_algo0300(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output;
 }
-float fm_algo0400(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0400(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -94,7 +94,7 @@ float fm_algo0400(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output;
 }
-float fm_algo0500(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0500(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -113,7 +113,7 @@ float fm_algo0500(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output;
 }
-float fm_algo0600(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0600(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -132,7 +132,7 @@ float fm_algo0600(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output;
 }
-float fm_algo0700(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0700(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -151,7 +151,7 @@ float fm_algo0700(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output;
 }
-float fm_algo0800(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0800(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -170,7 +170,7 @@ float fm_algo0800(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output;
 }
-float fm_algo0900(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo0900(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -189,7 +189,7 @@ float fm_algo0900(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output;
 }
-float fm_algo1000(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo1000(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -208,7 +208,7 @@ float fm_algo1000(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output + voice->op_states[2].output;
 }
-float fm_algo1100(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo1100(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
@@ -227,7 +227,7 @@ float fm_algo1100(fm_voice_t *voice, const float feedback_amp) {
     );
     return voice->op_states[0].output + voice->op_states[1].output + voice->op_states[2].output;
 }
-float fm_algo1200(fm_voice_t *voice, const float feedback_amp) {
+double fm_algo1200(fm_voice_t *voice, const float feedback_amp) {
     voice->op_states[3].output = fm_calc_op(
         voice->op_states[3].phase,
         voice->op_states[3].expression
